Don't delete prices if they're published again

+++ src/NHSD.GPIT.BuyingCatalogue.WebApp/Areas/Admin/Controllers/CatalogueSolutionListPriceController.cs
@@ -420,7 +420,14 @@ namespace NHSD.GPIT.BuyingCatalogue.WebApp.Areas.Admin.Controllers
         [HttpPost("{cataloguePriceId}/delete")]
         public async Task<IActionResult> DeleteListPrice(CatalogueItemId solutionId, int cataloguePriceId, DeleteItemConfirmationModel model)
         {
-            await solutionListPriceService.DeleteListPrice(solutionId, cataloguePriceId);
+            // TODO:
+            // In a scenario where a user unpublishes a price, opens the delete link in a second tab but first publishes the price again,
+            // Replace with an error page that explains why the price hasn't been deleted
+            var solution = await solutionListPriceService.GetSolutionWithListPrices(solutionId);
+            var price = solution.CataloguePrices.First(p => p.CataloguePriceId == cataloguePriceId);
+
+            if (price.PublishedStatus != PublicationStatus.Published)
+                await solutionListPriceService.DeleteListPrice(solutionId, cataloguePriceId);
 
             return RedirectToAction(nameof(Index), new { solutionId });
         }
@@ -444,6 +451,15 @@ namespace NHSD.GPIT.BuyingCatalogue.WebApp.Areas.Admin.Controllers
         [HttpPost("{cataloguePriceId}/tiers/{tierId}/delete")]
         public async Task<IActionResult> DeleteTieredPriceTier(CatalogueItemId solutionId, int cataloguePriceId, int tierId, DeleteItemConfirmationModel model, bool? isEditing = null)
         {
+            // TODO:
+            // In a scenario where a user unpublishes a price, opens the delete pricing tier link in a second tab but first publishes the price again,
+            // Replace with an error page that explains why the pricing tier hasn't been deleted
+            var solution = await solutionListPriceService.GetSolutionWithListPrices(solutionId);
+            var price = solution.CataloguePrices.First(p => p.CataloguePriceId == cataloguePriceId);
+
+            if (price.PublishedStatus == PublicationStatus.Published)
+                return RedirectToAction(nameof(EditTieredListPrice), new { solutionId, cataloguePriceId });
+
             await solutionListPriceService.DeletePriceTier(solutionId, cataloguePriceId, tierId);
 
             var actionName = GetAddOrEditBacklink(isEditing);
